+++ src/lib.rs
@@ -60,3 +60,6 @@
 pub mod parser;
 pub mod specification;
 pub mod translator;
+
+#[cfg(test)]
+mod tests;